rework the watch stream listener
Clear a resolved module only if its file is deleted. Indicate a
deleted file by setting its `id` property to 0.

Watch the "package.json" of any "node_modules" dependencies
instead of the root directory. This makes more sense, since
we were reading "package.json" on every change to begin with.
Also, reading "package.json" after a change is now triggered
by the parent package's watch stream (if a parent exists).

Change `readPackage` to return a boolean; true when the package
name and version have not changed. When "package.json" contains
malformed JSON, we are forgiving and avoid unloading the package.
Also, the `data` property of Package objects is now updated.

This commit relies on `Package` properties added in a later commit.