rework package unloading
Rename the `Package::_purge` method to `_unload` for clarity.

Add `users` property to the `Package` class, which is a `Set`
that tracks which packages are dependent on us.

Add `parent` property to the `Package` class, which is either
a package or null. The parent is determined by which package
required us first. The project root should be the only package
with a null parent.

Flesh out the `Package::_unload` method to perform the following
actions in the order they are listed:

  - update the `time` property of each file
  - update the `users` property of each dependency
  - unload packages where we are their parent
  - nullify our `files` cache
  - tell our `users` to delete us from their `files` caches
  - rebuild bundles that used us
  - delete ourself from the package cache
  - stop watching ourself