track packages with 1+ missed asset or package
When a package cannot be found, the dependent package has its
`missedPackage` property set to true. When its package.json is changed
or a package is added to node_modules, a rebuild is triggered.

When an asset cannot be found, the containing package has its
`missedAsset` property set to true. When an asset is added or renamed, a
rebuild is triggered.